1) Выведения методов mapRowToDirector и mapRowToMpa в отдельные классы 2) Использование метода batchUpdate, вместо цикла+update
