+++ src/main/java/ru/yandex/practicum/filmorate/mappers/DirectorRowMappers.java
@@ -1,0 +1,18 @@
+package ru.yandex.practicum.filmorate.mappers;
+
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.model.Director;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+@Component
+public class DirectorRowMappers {
+
+    public Director mapRowToDirector(ResultSet resultSet, int rowNum) throws SQLException {
+        return Director.builder()
+                .id(resultSet.getLong("id"))
+                .name(resultSet.getString("name"))
+                .build();
+    }
+}
