+++ src/main/java/ru/yandex/practicum/filmorate/mappers/MpaRowMappers.java
@@ -1,0 +1,19 @@
+package ru.yandex.practicum.filmorate.mappers;
+
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.model.Mpa;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+@Component
+public class MpaRowMappers {
+
+    public Mpa mapRowToMpa(ResultSet resultSet, int rowNum) throws SQLException {
+        return Mpa.builder()
+                .id(resultSet.getLong("id"))
+                .name(resultSet.getString("name"))
+                .build();
+    }
+
+}
